+++ lib/runners/harness.rb
@@ -56,11 +56,11 @@ module Runners
               case result
               when Results::Success
                 trace_writer.message "Removing issues from unchanged or untracked files..." do
-                  result.tap do |r|
-                    r.filter_issues(changes)
-                    r.add_git_blame_info(workspace)
-                  end
+                  result.filter_issues(changes)
+                  result.add_git_blame_info(workspace)
                 end
+                result.each_missing_id_warning { |msg| trace_writer.message(msg) }
+                result
               else
                 result
               end